Attempt to clean up ObstetricsVisit, ObstetricsData, and the controller.
Changed ObsetricsVisit to deal with the values for follow-up visits
instead of the initial pregnancy.

Changed ObstrictsData to use LocalDateTimes for Dates instead of
Strings. Changed ObstrictsData to use MySQL class instead of doing all
SQL functionality in controller.

Updated the controller to be able to use the new MySQL class.

Got rid of old JUnit tests for ObstetricsVisit.